Move bdd_policy and zdd_policy out of the 'public' API
