+++ src/adiar/bdd/apply.cpp
@@ -1,4 +1,5 @@
 #include <adiar/bdd.h>
+#include <adiar/bdd/bdd_policy.h>
 
 #include <adiar/file_stream.h>
 #include <adiar/file_writer.h>
