+++ src/adiar/bdd/bdd.cpp
@@ -11,6 +11,9 @@
 #include <adiar/internal/intercut.h>
 #include <adiar/internal/reduce.h>
 
+#include <adiar/bdd/bdd_policy.h>
+#include <adiar/zdd/zdd_policy.h>
+
 namespace adiar {
   //////////////////////////////////////////////////////////////////////////////
   // Constructors
