+++ src/adiar/zdd/pred.cpp
@@ -1,4 +1,5 @@
 #include <adiar/zdd.h>
+#include <adiar/zdd/zdd_policy.h>
 
 #include <adiar/internal/pred.h>
 #include <adiar/internal/product_construction.h>
